\name{linear_list}
\docType{data}
\alias{linear_list}
\title{Simulated data for a Gaussian 2-level HB regression model.}
\description{
  A list object containing data fixed and random effects as well group level data for prior regression. 
}
\usage{data(linear_list)}
\format{A list containing :
  \describe{
    \item{lower}{A data.frame with fixed and random effects data for the model.}
    \item{upper}{A data.frame with group level data for prior regression.}
  }
}
\source{
  Simulated data, generated from R code.
}

\keyword{datasets}
